Long

+++ Hahatun/Hahatun/Models/CertInfoModel.cs
@@ -7,11 +7,11 @@ namespace Hahatun.Models
 {
     public class CertInfoModel
     {
-        public int certnumber { get; set; }
+        public long certnumber { get; set; }
         public string certdate { get; set; }
         public int manufacturercode { get; set; }
         public string companycode { get; set; }
-        public int checkcode { get; set; }
+        public long checkcode { get; set; }
 
         public string productname { get; set; }
         public int status { get; set; }
